+++ bull-common/src/main/java/com/expediagroup/transformer/validator/Validator.java
@@ -20,7 +20,7 @@ import static java.util.Objects.isNull;
 import java.util.List;
 import java.util.Set;
 
-import javax.validation.ConstraintViolation;
+import jakarta.validation.ConstraintViolation;
 
 import com.expediagroup.transformer.error.InvalidBeanException;